moved examples to integration_tests, minor bug fixes
