+++ integration_tests/examples/audio.rs
@@ -8,12 +8,9 @@ use std::{
     time::Duration,
 };
 
-use crate::common::{
-    download_file, start_ffplay, start_websocket_thread, stream_audio, stream_video,
+use integration_tests::examples_common::{
+    self, download_file, start_ffplay, start_websocket_thread, stream_audio, stream_video,
 };
-
-#[path = "./common/common.rs"]
-mod common;
 
 const BUNNY_FILE_URL: &str =
     "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4";
@@ -59,7 +56,7 @@ fn start_example_client_code() -> Result<()> {
     let elephant_path = download_file(ELEPHANT_DREAM_FILE_URL, ELEPHANT_DREAM_FILE_PATH)?;
 
     info!("[example] Send register input request.");
-    common::post(
+    examples_common::post(
         "input/input_1/register",
         &json!({
             "type": "rtp_stream",
@@ -71,7 +68,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register input request.");
-    common::post(
+    examples_common::post(
         "input/input_2/register",
         &json!({
             "type": "rtp_stream",
@@ -83,7 +80,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register input request.");
-    common::post(
+    examples_common::post(
         "input/input_3/register",
         &json!({
             "type": "rtp_stream",
@@ -95,7 +92,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register input request.");
-    common::post(
+    examples_common::post(
         "input/input_4/register",
         &json!({
             "type": "rtp_stream",
@@ -107,7 +104,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register output request.");
-    common::post(
+    examples_common::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -143,7 +140,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register output request.");
-    common::post(
+    examples_common::post(
         "output/output_2/register",
         &json!({
             "type": "rtp_stream",
@@ -167,7 +164,7 @@ fn start_example_client_code() -> Result<()> {
     std::thread::sleep(Duration::from_millis(500));
 
     info!("[example] Start pipeline");
-    common::post("start", &json!({}))?;
+    examples_common::post("start", &json!({}))?;
 
     stream_video(IP, INPUT_1_PORT, bunny_path.clone())?;
     stream_audio(IP, INPUT_2_PORT, bunny_path, "libopus")?;
