+++ integration_tests/examples/decklink.rs
@@ -8,10 +8,7 @@ use std::{
     time::Duration,
 };
 
-use crate::common::start_websocket_thread;
-
-#[path = "./common/common.rs"]
-mod common;
+use integration_tests::examples_common::{self, start_websocket_thread};
 
 const VIDEO_RESOLUTION: Resolution = Resolution {
     width: 1280,
@@ -39,7 +36,7 @@ fn start_example_client_code() -> Result<()> {
     start_websocket_thread();
 
     info!("[example] Send register input request.");
-    common::post(
+    examples_common::post(
         "input/input_1/register",
         &json!({
             "type": "decklink",
@@ -49,7 +46,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register output video request.");
-    common::post(
+    examples_common::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -116,7 +113,7 @@ fn start_example_client_code() -> Result<()> {
     std::thread::sleep(Duration::from_millis(1000));
 
     info!("[example] Start pipeline");
-    common::post("start", &json!({}))?;
+    examples_common::post("start", &json!({}))?;
 
     Ok(())
 }
