+++ integration_tests/examples/silly.wgsl
