+++ integration_tests/examples/transition.rs
@@ -4,10 +4,9 @@ use log::{error, info};
 use serde_json::json;
 use std::{thread, time::Duration};
 
-use crate::common::{start_ffplay, start_websocket_thread, stream_ffmpeg_testsrc};
-
-#[path = "./common/common.rs"]
-mod common;
+use integration_tests::examples_common::{
+    self, start_ffplay, start_websocket_thread, stream_ffmpeg_testsrc,
+};
 
 const VIDEO_RESOLUTION: Resolution = Resolution {
     width: 1920,
@@ -36,7 +35,7 @@ fn start_example_client_code() -> Result<()> {
     start_websocket_thread();
 
     info!("[example] Send register input request.");
-    common::post(
+    examples_common::post(
         "input/input_1/register",
         &json!({
             "type": "rtp_stream",
@@ -49,7 +48,7 @@ fn start_example_client_code() -> Result<()> {
 
     let shader_source = include_str!("./silly.wgsl");
     info!("[example] Register shader transform");
-    common::post(
+    examples_common::post(
         "shader/example_shader/register",
         &json!({
             "source": shader_source,
@@ -57,7 +56,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Register static image");
-    common::post(
+    examples_common::post(
         "image/example_image/register",
         &json!({
             "asset_type": "gif",
@@ -158,7 +157,7 @@ fn start_example_client_code() -> Result<()> {
     });
 
     info!("[example] Send register output request.");
-    common::post(
+    examples_common::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -181,7 +180,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Start pipeline");
-    common::post("start", &json!({}))?;
+    examples_common::post("start", &json!({}))?;
 
     info!("[example] Start input stream");
     stream_ffmpeg_testsrc(IP, INPUT_PORT, VIDEO_RESOLUTION)?;
@@ -189,7 +188,7 @@ fn start_example_client_code() -> Result<()> {
     thread::sleep(Duration::from_secs(5));
 
     info!("[example] Update output");
-    common::post(
+    examples_common::post(
         "output/output_1/update",
         &json!({
             "video": {
@@ -201,7 +200,7 @@ fn start_example_client_code() -> Result<()> {
     thread::sleep(Duration::from_secs(2));
 
     info!("[example] Update output");
-    common::post(
+    examples_common::post(
         "output/output_1/update",
         &json!({
             "video": {
