+++ integration_tests/examples/web_view.rs
@@ -7,10 +7,7 @@ use std::{
     thread::{self},
 };
 
-use crate::common::{download_file, start_ffplay, stream_video};
-
-#[path = "./common/common.rs"]
-mod common;
+use integration_tests::examples_common::{self, download_file, start_ffplay, stream_video};
 
 const SAMPLE_FILE_URL: &str = "https://filesamples.com/samples/video/mp4/sample_1280x720.mp4";
 const SAMPLE_FILE_PATH: &str = "examples/assets/sample_1280_720.mp4";
@@ -64,7 +61,7 @@ fn start_example_client_code() -> Result<()> {
         .to_string();
 
     info!("[example] Send register input request.");
-    common::post(
+    examples_common::post(
         "input/input_1/register",
         &json!({
             "type": "rtp_stream",
@@ -76,7 +73,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Register web renderer transform");
-    common::post(
+    examples_common::post(
         "web-renderer/example_website/register",
         &json!({
             "url": format!("file://{html_file_path}"), // or other way of providing source
@@ -85,7 +82,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register output request.");
-    common::post(
+    examples_common::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -119,7 +116,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Start pipeline");
-    common::post("start", &json!({}))?;
+    examples_common::post("start", &json!({}))?;
 
     stream_video(IP, INPUT_PORT, sample_path)?;
     Ok(())
